feat: CRUD items, Add favicon and improve user feedback in forms
- Added favicon to login and registration pages for branding.
- Updated input fields in login and registration forms to include autocomplete attributes for better user experience.
- Implemented success and error message notifications in the layout component for improved user feedback.
- Refactored navbar to include favicon and reorganized button layout for better accessibility.
- Removed redundant error handling from category and user creation/edit pages, utilizing the new notification system instead.
- Created item management views including create, edit, delete, and index with improved UI and UX.
- Enhanced item forms with better input handling and image upload previews.

+++ database/migrations/2026_01_02_132827_create_rentals_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->date('start_date');
             $table->date('end_date');
-            $table->decimal('total_price', 12, 2);
+            $table->integer('total_price');
             $table->enum('status', [
                 'pending',
                 'paid',
